author showed but frontend cant add or update

+++ ProjectBackend/src/main/java/org/isdb/ProjectBackend/repository/AuthorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Integer> {
-    // Additional query methods can go here (if needed)
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+	// Additional query methods can go here (if needed)
 }
